MAINTENANCE: Remove DPKG specific variables used for non-standard root
DPKG_ROOT was a tool used for access to chroot filesystem in connection to the 'gcc --sysroot'. The build flow no longer uses a chroot (not in Docker container and not in actual native system), so this option is no longer needed and would only cause a confusion to newly minted Machinekit developers.

Users who need this functionality for some reason or other will be better equipped to add it add ad-hoc.